# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HandyMan API
  version: "1.0.0"
  description: >
    Public API for the HandyMan marketplace. Additive-only forever (CLAUDE.md rule #4): never
    remove a field, never tighten a rule, never change an enum's meaning. This spec is the
    SOURCE OF TRUTH for the TypeScript client — the client is generated from it, never hand-written.
servers:
  - url: /api/v1
    description: Version 1 (URL-versioned)
...
paths:
...
  /provider/payouts:
    post:
      tags: [Platform]
      operationId: requestPayout
      summary: Request a payout of the provider's payable balance
      description: >
        Withdraw available payable funds to a mobile-money wallet. Idempotent on the Idempotency-Key.
        Funds are reserved by the pending payout; the ledger posts only on gateway confirmation.
        Rejected (422) if the requested amount exceeds the available (unreserved) payable balance.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_minor, msisdn]
              properties:
                amount_minor: { type: integer, minimum: 1 }
                msisdn: { type: string, maxLength: 20 }
                method:
                  type: string
                  enum: [mtn_momo, orange_money]
                  description: The rail to pay out on. Optional — absent, the number's prefix decides; a number no operator claims is refused (422 unknown-mobile-rail).
      responses:
        "201": { description: The requested payout, content: { application/json: { schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Payout" } } } } } }
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
...
  /payment-intents:
    post:
      tags: [Platform]
      operationId: initiatePaymentIntent
      summary: Start a MoMo collection (escrow or lead credits)
      description: >
        Initiates a mobile-money collection. The response rests in `processing` until the payer
        answers the USSD prompt — the client shows a countdown to `expires_at` and (for redirect
        flows) follows `payment_url`. Idempotent on the Idempotency-Key header: the same key returns
        the same intent, never a second charge. Resolution arrives via webhook or reconciliation poll.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [purpose, amount_minor, msisdn]
              properties:
                purpose: { type: string, enum: [escrow, lead_credits] }
                amount_minor: { type: integer, minimum: 1 }
                msisdn: { type: string, maxLength: 20 }
                method:
                  type: string
                  enum: [mtn_momo, orange_money]
                  description: The rail to prompt the payer on. Optional — absent, the number's prefix decides; a number no operator claims is refused (422 unknown-mobile-rail). Cash is never a collection.
                engagement_id: { type: [string, "null"], format: uuid, description: "Required when purpose is escrow." }
      responses:
        "201": { description: The pending payment intent, content: { application/json: { schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/PaymentIntent" } } } } } }
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: A client-generated UUID. Replaying it returns the stored response (CLAUDE.md rule #3).
      schema:
        type: string
        format: uuid
...
  schemas:
    Money:
      type: object
      description: Money is ALWAYS integer minor units + explicit currency. Never a float, never a formatted string.
      required: [amount_minor, currency]
      properties:
        amount_minor:
          type: integer
          format: int64
        currency:
          type: string
          minLength: 3
          maxLength: 3
          examples: ["XAF"]
...
    PaymentMethod:
      type: string
      description: >-
        mtn_momo and orange_money are the two mobile rails a collection or payout goes out on; cash is
        recorded after the fact as a cash settlement and never touches the gateway.
      enum: [mtn_momo, orange_money, cash]
    Payout:
      type: object
      required: [id, amount, msisdn, status, requested_at]
      properties:
        id: { type: string, format: uuid }
        amount: { $ref: "#/components/schemas/Money" }
        msisdn: { type: string }
        method:
          description: The rail the money goes out on; null only on rows older than the field.
          oneOf:
            - $ref: "#/components/schemas/PaymentMethod"
            - type: "null"
        status: { type: string, enum: [pending, processing, succeeded, failed, expired] }
        external_ref: { type: [string, "null"] }
        requested_at: { type: string, format: date-time }
        resolved_at: { type: [string, "null"], format: date-time }
        reversed_at: { type: [string, "null"], format: date-time }
        failure_code: { type: [string, "null"] }
    PaymentIntent:
      type: object
      required: [id, purpose, status, amount, msisdn, expires_at, initiated_at]
      properties:
        id: { type: string, format: uuid }
        purpose: { type: string, enum: [escrow, lead_credits] }
        status: { type: string, enum: [pending, processing, succeeded, failed, expired] }
        amount: { $ref: "#/components/schemas/Money" }
        msisdn: { type: string }
        method:
          description: The rail the payer is prompted on; null only on rows older than the field.
          oneOf:
            - $ref: "#/components/schemas/PaymentMethod"
            - type: "null"
        external_ref: { type: [string, "null"] }
        payment_url: { type: [string, "null"] }
        expires_at: { type: string, format: date-time }
        initiated_at: { type: string, format: date-time }
        resolved_at: { type: [string, "null"], format: date-time }
        failure_code: { type: [string, "null"] }
...
    Problem:
      type: object
      description: RFC 7807 problem+json — the single error shape for the whole API.
      required: [type, title, status, detail, trace_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        trace_id:
          type: string
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
  responses:
    Problem:
      description: An error (RFC 7807 problem+json)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationProblem:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"
...
security:
  - bearerAuth: []